+++ guillotina/asgi.py
@@ -10,6 +10,7 @@ from guillotina.middlewares import ErrorsMiddleware
 from guillotina.request import Request
 from guillotina.response import Response
 from guillotina.traversal import apply_rendering
+from guillotina.utils import get_dotted_name
 from guillotina.utils import resolve_dotted_name
 
 import asyncio
@@ -179,6 +180,10 @@ class Guillotina:
     async def request_handler(self, request, retries=0):
         try:
             route = await self.router.resolve(request)
+            # The key 'endpoint' in scope is used by sentry-sdk to display
+            # the name of the failed view
+            if hasattr(route, "view"):
+                request.scope["endpoint"] = Endpoint(route.view)
             resp = await route.handler(request)
             return resp
 
@@ -211,3 +216,9 @@ class Guillotina:
                     )
                 )
                 raise HTTPConflict()
+
+
+class Endpoint:
+    def __init__(self, view):
+        self.__module__ = view.__module__
+        self.__qualname__ = get_dotted_name(view).split(".")[-1]